Seperated LoggedIn and Login navigation
