+++ Projects/App/Sources/LoggedInNavigation.swift
@@ -1,0 +1,185 @@
+import Chat
+import Claims
+import Contracts
+import EditCoInsured
+import EditCoInsuredShared
+import Forever
+import Foundation
+import Home
+import MoveFlow
+import Payment
+import Presentation
+import Profile
+import SwiftUI
+import hCore
+import hCoreUI
+
+struct LoggedInNavigation: View {
+    @StateObject private var vm = LoggedInNavigationViewModel()
+    @StateObject private var homeNavigationVm = HomeNavigationViewModel()
+    @StateObject var router = Router()
+    var body: some View {
+        TabView(selection: $vm.selectedTab) {
+            Group {
+                homeTab
+                    .environmentObject(router)
+                contractsTab
+
+                let store: ContractStore = globalPresentableStoreContainer.get()
+                if !store.state.activeContracts.allSatisfy({ $0.isNonPayingMember })
+                    || store.state.activeContracts.isEmpty
+                {
+                    foreverTab
+                }
+
+                if Dependencies.featureFlags().isPaymentScreenEnabled {
+                    paymentsTab
+                }
+                profileTab
+            }
+        }
+        .tint(hTextColor.primary)
+    }
+
+    var homeTab: some View {
+        let claims = Claims()
+
+        return RouterHost(router: router) {
+            HomeView(
+                claimsContent: claims,
+                memberId: {
+                    let profileStrore: ProfileStore = globalPresentableStoreContainer.get()
+                    return profileStrore.state.memberDetails?.id ?? ""
+                }
+            )
+            .routerDestination(for: ClaimModel.self) { claim in
+                ClaimDetailView(claim: claim)
+                    .environmentObject(homeNavigationVm)
+            }
+        }
+        .environmentObject(homeNavigationVm)
+        .detent(
+            presented: $homeNavigationVm.isSubmitClaimPresented,
+            style: .height
+        ) {
+            HonestyPledge(onConfirmAction: {})
+                .embededInNavigation()
+        }
+        .detent(
+            item: $homeNavigationVm.document,
+            style: .large
+        ) { document in
+            if let url = URL(string: document.url) {
+                PDFPreview(document: .init(url: url, title: document.displayName))
+                    .embededInNavigation(options: [.navigationType(type: .large)])
+            }
+        }
+        .detent(
+            presented: $homeNavigationVm.navBarItems.isFirstVetPresented,
+            style: .height
+        ) {
+            let store: HomeStore = globalPresentableStoreContainer.get()
+            return FirstVetView(partners: store.state.quickActions.getFirstVetPartners ?? [])
+        }
+        .detent(
+            presented: $homeNavigationVm.navBarItems.isNewOfferPresented,
+            style: .height
+        ) {
+            CrossSellingScreen()
+        }
+        .detent(
+            presented: $homeNavigationVm.isCoInsuredPresented,
+            style: .height
+        ) {
+            let contractStore: ContractStore = globalPresentableStoreContainer.get()
+            let contractsSupportingCoInsured = contractStore.state.activeContracts
+                .filter({ $0.showEditCoInsuredInfo })
+                .compactMap({
+                    InsuredPeopleConfig(contract: $0, fromInfoCard: true)
+                })
+            EditCoInsuredViewJourney(configs: contractsSupportingCoInsured)
+        }
+        .fullScreenCover(isPresented: $homeNavigationVm.isHelpCenterPresented) {
+            HelpCenterNavigation()
+                .environmentObject(homeNavigationVm)
+        }
+
+        .detent(
+            item: $homeNavigationVm.openChat,
+            style: .large,
+            options: $homeNavigationVm.openChatOptions,
+            content: { openChat in
+                ChatScreen(vm: .init(topicType: openChat.topic))
+                    .navigationTitle(L10n.chatTitle)
+                    .embededInNavigation()
+            }
+        )
+        .tabItem {
+            Image(uiImage: vm.selectedTab == 0 ? hCoreUIAssets.homeTabActive.image : hCoreUIAssets.homeTab.image)
+            hText(L10n.tabHomeTitle)
+        }
+        .tag(0)
+    }
+
+    var contractsTab: some View {
+        ContractsNavigation { redirectType in
+            switch redirectType {
+            case let .editCoInsured(editCoInsuredConfig):
+                EditCoInsuredViewJourney(configs: [editCoInsuredConfig])
+            case .chat:
+                ChatScreen(vm: .init(topicType: nil))
+                    .presentationDetents([.large, .medium])
+            case .movingFlow:
+                MovingFlowViewJourney()
+            case let .pdf(document):
+                PDFPreview(document: .init(url: document.url, title: document.title))
+            }
+        }
+        .tabItem {
+            Image(
+                uiImage: vm.selectedTab == 1
+                    ? hCoreUIAssets.contractTabActive.image : hCoreUIAssets.contractTab.image
+            )
+            hText(L10n.tabInsurancesTitle)
+        }
+        .tag(1)
+    }
+
+    var foreverTab: some View {
+        ForeverView()
+            .tabItem {
+                Image(
+                    uiImage: vm.selectedTab == 2 ? hCoreUIAssets.foreverTabActive.image : hCoreUIAssets.foreverTab.image
+                )
+                hText(L10n.tabReferralsTitle)
+            }
+            .tag(2)
+    }
+
+    var paymentsTab: some View {
+        PaymentsView()
+            .tabItem {
+                Image(
+                    uiImage: vm.selectedTab == 3
+                        ? hCoreUIAssets.paymentsTabActive.image : hCoreUIAssets.paymentsTab.image
+                )
+                hText(L10n.tabPaymentsTitle)
+            }
+            .tag(3)
+    }
+
+    var profileTab: some View {
+        ProfileView()
+            .tabItem {
+                Image(
+                    uiImage: vm.selectedTab == 4 ? hCoreUIAssets.profileTabActive.image : hCoreUIAssets.profileTab.image
+                )
+                hText(L10n.ProfileTab.title)
+            }
+            .tag(4)
+    }
+}
+
+class LoggedInNavigationViewModel: ObservableObject {
+    @Published var selectedTab = 0
+}
