+++ Projects/Profile/Sources/Views/Screens/Profile/Profile.swift
@@ -27,6 +27,7 @@ public struct ProfileView: View {
             primaryButton: .cancel(Text(L10n.logoutAlertActionCancel)),
             secondaryButton: .destructive(Text(L10n.logoutAlertActionConfirm)) {
                 ApplicationState.preserveState(.notLoggedIn)
+                ApplicationState.state = .notLoggedIn
                 store.send(.logout)
             }
         )
