+++ Projects/hCore/Sources/ApplicationState.swift
@@ -1,7 +1,9 @@
 import Foundation
+import SwiftUI
 import hGraphQL
 
 public struct ApplicationState {
+    @AppStorage(key) public static var state: ApplicationState.Screen = .notLoggedIn
     public enum Screen: String {
         case onboardingChat, offer, loggedIn, languagePicker, notLoggedIn, onboarding, impersonation
 
@@ -91,3 +93,26 @@ public struct ApplicationState {
         return .en_SE
     }
 }
+//
+//import Foundation
+//import Combine
+//
+//@propertyWrapper
+//public struct UserDefault<T> {
+//    let key: String
+//    let defaultValue: T
+//
+//    init(_ key: String, defaultValue: T) {
+//        self.key = key
+//        self.defaultValue = defaultValue
+//    }
+//
+//    public var wrappedValue: T {
+//        get {
+//            return UserDefaults.standard.object(forKey: key) as? T ?? defaultValue
+//        }
+//        set {
+//            UserDefaults.standard.set(newValue, forKey: key)
+//        }
+//    }
+//}
